# openapi.yaml
openapi: 3.0.3
info:
  title: Forum API
  version: 1.0.0
  description: REST API для форумной платформы (посты, комментарии, лайки, пользователи, поиск, модерация, админка).

servers:
  - url: http://localhost:8080

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    RegisterRequest:
      type: object
      properties:
        email: { type: string }
        username: { type: string }
        password: { type: string }

    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }

    LoginResponse:
      type: object
      properties:
        token: { type: string }

    PostResponse:
      type: object
      properties:
        id: { type: string }
        authorId: { type: string }
        title: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
        editedAt: { type: string, format: date-time, nullable: true }
        categoryId: { type: string }
        likes: { type: integer }
        likedByMe: { type: boolean }

    CreatePostRequest:
      type: object
      properties:
        title: { type: string }
        content: { type: string }
        categoryId: { type: string }

    UpdatePostRequest:
      type: object
      properties:
        title: { type: string }
        content: { type: string }
        categoryId: { type: string, nullable: true }

    CommentResponse:
      type: object
      properties:
        id: { type: string }
        authorId: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
        editedAt: { type: string, format: date-time, nullable: true }
        likes: { type: integer }
        likedByMe: { type: boolean }

    CreateCommentRequest:
      type: object
      properties:
        content: { type: string }

    UpdateCommentRequest:
      type: object
      properties:
        content: { type: string }

    UserResponse:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        username: { type: string }
        avatarId: { type: string, nullable: true }
        role: { type: string }

    PublicUserResponse:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        avatarId: { type: string, nullable: true }

    UpdateUserRequest:
      type: object
      properties:
        email: { type: string, nullable: true }
        username: { type: string, nullable: true }
        newPassword: { type: string, nullable: true }

    ChangeRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [ADMIN, MODERATOR, USER]

paths:

  #########################################
  # AUTH
  #########################################

  /api/auth/register:
    post:
      summary: Регистрация
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }

  /api/auth/login:
    post:
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }

  #########################################
  # POSTS
  #########################################

  /api/posts:
    get:
      summary: Получить список постов
      security: []
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: size, in: query, schema: { type: integer } }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PostResponse" }

    post:
      summary: Создать пост
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePostRequest" }
      responses:
        200:
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostResponse" }

  /api/posts/{id}:
    get:
      summary: Получить пост
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostResponse" }

    put:
      summary: Обновить пост
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePostRequest" }
      responses:
        200:
          description: OK

    delete:
      summary: Удалить пост
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true }
      responses:
        204:
          description: No Content

  /api/posts/category/{slug}:
    get:
      summary: Посты категории
      parameters:
        - { name: slug, in: path, required: true }
        - { name: page, in: query, schema: { type: integer } }
        - { name: size, in: query, schema: { type: integer } }
      responses:
        200:
          description: OK

  #########################################
  # COMMENTS
  #########################################

  /api/posts/{postId}/comments:
    get:
      summary: Получить комментарии поста
      parameters:
        - { name: postId, in: path, required: true }
        - { name: page, in: query }
        - { name: size, in: query }
      responses:
        200:
          description: OK

    post:
      summary: Создать комментарий
      security:
        - bearerAuth: []
      parameters:
        - { name: postId, in: path, required: true }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommentRequest" }
      responses:
        200:
          description: OK

  /api/comments/{id}:
    put:
      summary: Обновить комментарий
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCommentRequest" }
      responses:
        200:
          description: OK

    delete:
      summary: Удалить комментарий
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true }
      responses:
        204:
          description: No Content

  #########################################
  # LIKES
  #########################################

  /api/posts/{postId}/like:
    post:
      summary: Лайк поста
      security:
        - bearerAuth: []
      parameters:
        - { name: postId, in: path, required: true }
      responses:
        200: { description: OK }

    delete:
      summary: Удалить лайк с поста
      security:
        - bearerAuth: []
      parameters:
        - { name: postId, in: path, required: true }
      responses:
        200: { description: OK }

  /api/comments/{commentId}/like:
    post:
      summary: Лайк комментария
      security:
        - bearerAuth: []
      parameters:
        - { name: commentId, in: path }
      responses:
        200: { description: OK }

    delete:
      summary: Удалить лайк комментария
      security:
        - bearerAuth: []
      responses:
        200: { description: OK }

  #########################################
  # USERS
  #########################################

  /api/user/me:
    get:
      summary: Текущий пользователь
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

    put:
      summary: Изменить профиль
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        200: { description: OK }

  /api/user/{id}:
    get:
      summary: Получить публичный профиль
      security: []
      parameters:
        - { name: id, in: path, required: true }
      responses:
        200:
          description: OK

  #########################################
  # AVATAR
  #########################################

  /api/user/me/avatar:
    post:
      summary: Загрузить аватар
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200: { description: OK }

    delete:
      summary: Удалить аватар
      security:
        - bearerAuth: []
      responses:
        204: { description: No Content }

  #########################################
  # SEARCH
  #########################################

  /api/search/posts:
    get:
      summary: Поиск постов
      parameters:
        - { name: query, in: query }
        - { name: page, in: query }
        - { name: size, in: query }
      responses:
        200:
          description: OK

  /api/search/users:
    get:
      summary: Поиск пользователей
      parameters:
        - { name: query, in: query }
      responses:
        200:
          description: OK

  #########################################
  # ADMIN
  #########################################

  /api/admin/posts/{id}/pin:
    post:
      summary: Закрепить пост
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path }
      responses:
        200: { description: OK }

  /api/admin/posts/{id}/lock:
    post:
      summary: Закрыть пост
      security:
        - bearerAuth: []
      responses:
        200: { description: OK }

  /api/admin/users/{id}:
    delete:
      summary: Забанить пользователя
      security:
        - bearerAuth: []
      responses:
        200: { description: OK }

  /api/admin/users/{id}/role:
    put:
      summary: Смена роли пользователя
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeRoleRequest" }
      responses:
        200:
          description: OK

  #########################################
  # MODERATOR
  #########################################

  /api/mod/posts/{id}/pin:
    post:
      summary: Модератор — закрепить пост
      security:
        - bearerAuth: []
      responses:
        200: { description: OK }

  /api/mod/posts/{id}/lock:
    post:
      summary: Модератор — закрыть пост
      security:
        - bearerAuth: []
      responses:
        200: { description: OK }

  /api/mod/posts/{id}:
    delete:
      summary: Модератор — удалить пост
      security:
        - bearerAuth: []
      responses:
        200: { description: OK }

  /api/mod/comments/{id}:
    delete:
      summary: Модератор — удалить комментарий
      security:
        - bearerAuth: []
      responses:
        200: { description: OK }